requirements.txt
streamlit
ultralytics
opencv-python
numpy
pandas
python-dotenv
tensorflow